Update and enhance cookiecutter, docs, docker, etc.

+++ hooks/post_gen_project.py
@@ -32,10 +32,7 @@ DEBUG_VALUE = "debug"
 
 
 def remove_open_source_files():
-    file_names = [
-        "CONTRIBUTORS.txt",
-        "LICENSE",
-    ]
+    file_names = ["CONTRIBUTORS.txt", "LICENSE"]
     for file_name in file_names:
         os.remove(file_name)
 
@@ -64,9 +61,19 @@ def remove_docker_files():
         os.remove(file_name)
 
 
+def remove_utility_files():
+    shutil.rmtree("utility")
+
+
 def remove_heroku_files():
     file_names = ["Procfile", "runtime.txt", "requirements.txt"]
     for file_name in file_names:
+        if (
+            file_name == "requirements.txt"
+            and "{{ cookiecutter.use_travisci }}".lower() == "y"
+        ):
+            # don't remove the file if we are using travisci but not using heroku
+            continue
         os.remove(file_name)
 
 
@@ -168,11 +175,7 @@ def generate_postgres_user(debug=False):
 
 
 def set_postgres_user(file_path, value):
-    postgres_user = set_flag(
-        file_path,
-        "!!!SET POSTGRES_USER!!!",
-        value=value,
-    )
+    postgres_user = set_flag(file_path, "!!!SET POSTGRES_USER!!!", value=value)
     return postgres_user
 
 
@@ -190,9 +193,7 @@ def set_postgres_password(file_path, value=None):
 
 def set_celery_flower_user(file_path, value):
     celery_flower_user = set_flag(
-        file_path,
-        "!!!SET CELERY_FLOWER_USER!!!",
-        value=value,
+        file_path, "!!!SET CELERY_FLOWER_USER!!!", value=value
     )
     return celery_flower_user
 
@@ -215,11 +216,7 @@ def append_to_gitignore_file(s):
         gitignore_file.write(os.linesep)
 
 
-def set_flags_in_envs(
-    postgres_user,
-    celery_flower_user,
-    debug=False,
-):
+def set_flags_in_envs(postgres_user, celery_flower_user, debug=False):
     local_django_envs_path = os.path.join(".envs", ".local", ".django")
     production_django_envs_path = os.path.join(".envs", ".production", ".django")
     local_postgres_envs_path = os.path.join(".envs", ".local", ".postgres")
@@ -229,14 +226,22 @@ def set_flags_in_envs(
     set_django_admin_url(production_django_envs_path)
 
     set_postgres_user(local_postgres_envs_path, value=postgres_user)
-    set_postgres_password(local_postgres_envs_path, value=DEBUG_VALUE if debug else None)
+    set_postgres_password(
+        local_postgres_envs_path, value=DEBUG_VALUE if debug else None
+    )
     set_postgres_user(production_postgres_envs_path, value=postgres_user)
-    set_postgres_password(production_postgres_envs_path, value=DEBUG_VALUE if debug else None)
+    set_postgres_password(
+        production_postgres_envs_path, value=DEBUG_VALUE if debug else None
+    )
 
     set_celery_flower_user(local_django_envs_path, value=celery_flower_user)
-    set_celery_flower_password(local_django_envs_path, value=DEBUG_VALUE if debug else None)
+    set_celery_flower_password(
+        local_django_envs_path, value=DEBUG_VALUE if debug else None
+    )
     set_celery_flower_user(production_django_envs_path, value=celery_flower_user)
-    set_celery_flower_password(production_django_envs_path, value=DEBUG_VALUE if debug else None)
+    set_celery_flower_password(
+        production_django_envs_path, value=DEBUG_VALUE if debug else None
+    )
 
 
 def set_flags_in_settings_files():
@@ -246,6 +251,7 @@ def set_flags_in_settings_files():
 
 def remove_envs_and_associated_files():
     shutil.rmtree(".envs")
+    os.remove("merge_production_dotenvs_in_dotenv.py")
 
 
 def remove_celery_compose_dirs():
@@ -271,7 +277,9 @@ def main():
     if "{{ cookiecutter.use_pycharm }}".lower() == "n":
         remove_pycharm_files()
 
-    if "{{ cookiecutter.use_docker }}".lower() == "n":
+    if "{{ cookiecutter.use_docker }}".lower() == "y":
+        remove_utility_files()
+    else:
         remove_docker_files()
 
     if "{{ cookiecutter.use_heroku }}".lower() == "n":
@@ -284,8 +292,8 @@ def main():
         if "{{ cookiecutter.keep_local_envs_in_vcs }}".lower() == "y":
             print(
                 INFO + ".env(s) are only utilized when Docker Compose and/or "
-                       "Heroku support is enabled so keeping them does not "
-                       "make sense given your current setup." + TERMINATOR
+                "Heroku support is enabled so keeping them does not "
+                "make sense given your current setup." + TERMINATOR
             )
         remove_envs_and_associated_files()
     else:
